compute_prompts: update the bill similarity computation to use a real bill-user matrix, which seems to produce better results

+++ site/popvox/db/compute_prompts.py
@@ -5,41 +5,34 @@
 import math
 import itertools
 
-from popvox.models import UserComment, BillSimilarity
+from django.db.models import Count
 
-# Loop through the comments and collect in a list all of the bills commented on by
-# each user, user by user. Then increment a cooccurrence matrix for each pair
-# in the bill list. The traditional thing to do would be to build a bill-user matrix,
-# but since bills come from a smaller set than users we'll build a bill-bill matrix.
+from popvox.models import Bill, UserComment, BillSimilarity
+from popvox.govtrack import CURRENT_CONGRESS
+
+# Create a bill-user matrix, using our top N users.
+
+num_bills = Bill.objects.all().count()
 matrix = { }
-bills = []
-last_user = None
-for c in itertools.chain(UserComment.objects.only("id", "user", "bill").order_by('user', '-message', '-created').iterator(), [None]):
-	if c == None or c.user_id != last_user:
-		# bills contains all of the bills commented on by the last user
-		for b1 in bills:
-			if not b1 in matrix: matrix[b1] = {}
-			for b2 in bills:
-				if not b2 in matrix[b1]: matrix[b1][b2] = 0 
-				matrix[b1][b2] += 1.0/float(len(bills))
-		
-		# reset and continue with the next user
-		bills = []
-		if c == None: break # finished
-		
-	last_user = c.user_id
-	
-	# only take the first 100 comments from a user, looking first at the comments
-	# they left messages on, and then the most recent.
-	if len(bills) < 100:
-		bills.append(c.bill_id)
+bill_is_current = set()
+num_users = 0
+for rec in UserComment.objects.values("user").annotate(count=Count("user")).filter(count__lt=num_bills/100).order_by('-count')[0:num_bills]:
+	user = rec["user"]
+	for billid, billcongress in UserComment.objects.filter(user=user).values_list("bill", "bill__congressnumber"):
+		if not billid in matrix: matrix[billid] = set()
+		matrix[billid].add(user)
+		if billcongress == CURRENT_CONGRESS: bill_is_current.add(billid)
+	num_users += 1
+
+#print num_bills, num_users
 
 def dot(v1, v2):
-	# Compute dot product of two vectors, each stored in a sparse form as a dict.
+	# Compute dot product of two binary vectors, each stored in a sparse form as a set.
+	if v1 == v2: return len(v1)
 	ret = 0.0
 	for k in v1:
 		if k in v2:
-			ret += v1[k] * v2[k]
+			ret += 1
 	return ret
 
 def cosine(v1, v2):
@@ -47,29 +40,32 @@ def cosine(v1, v2):
 	# in a sparse form as a dict.
 	return dot(v1, v2) / math.sqrt(dot(v1,v1) * dot(v2,v2))
 
-# Find the max value in the cooccurrence matrix.
-max_val = None
-for b1 in matrix:
-	for b2 in matrix[b1]:
-		max_val = max(max_val, matrix[b1][b2])
-
 # Compute cosine similarity between each row.
 for b1 in matrix:
 	# Compute similarity scores between this bill and all bills with a greater id.
+	if len(matrix[b1]) < 5: continue
 	sim = {}
-	for b2 in matrix[b1]: # limit the scoring to bills that at least one User commented on both
-		if b1 < b2: # for any pair, we only need to do similarity score once
-			sim[b2] = cosine(matrix[b1], matrix[b2])
-			
-			# weight the score by the highest cooccurrence rate in each, if the cooccurrence
-			# rate is very low.
-			v = max(matrix[b1].values())/max_val * max(matrix[b2].values())/max_val
-			if v < .01:
-				sim[b2] *= 100.0*v
-	
+	b1_is_current = (b1 in bill_is_current)
+	for b2 in matrix:
+		if b1 >= b2: continue # for any pair, we only need to do similarity score once
+		if not b1_is_current and b2 not in bill_is_current: continue # at least one must be current
+		
+		if len(matrix[b2]) < 5: continue
+		
+		v = cosine(matrix[b1], matrix[b2])
+		if v < .66: continue # not similar enough to matter
+		sim[b2] = v
+
 	# Sort the similarity scores in descending order.
 	sim = list(sim.items()) # make it a (k,v) list
 	sim.sort(key = lambda x : -x[1])
+	
+	if len(sim) == 0:
+		continue
+		
+	#print Bill.objects.get(id=b1).title.encode("utf8")
+	#print Bill.objects.get(id=sim[0][0]).title.encode("utf8")
+	#print
 	
 	# Store the top scores.
 	BillSimilarity.objects.filter(bill1 = b1).delete()
